add response to server for TransactionsHistory and Expenses
