+++ app/build.gradle.kts
@@ -5,6 +5,7 @@ plugins {
     alias(libs.plugins.kotlin.android)
     alias(libs.plugins.kotlin.compose)
     alias(libs.plugins.kotlin.serialization)
+    //alias(libs.plugins.ksp)
 }
 
 android {
@@ -71,6 +72,10 @@ dependencies {
     implementation(libs.retrofit)
     implementation(libs.retrofit2.converter.kotlinx.serialization)
     implementation(libs.kotlinx.serialization.json)
+//    implementation(libs.hilt.android)
+//    implementation(libs.mediation.test.suite)
+//    ksp(libs.dagger.compiler)
+//    ksp(libs.hilt.android.compiler)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
