+++ app/src/main/java/com/example/myfinance/app/MainActivity.kt
@@ -20,12 +20,6 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
 
-//        lifecycleScope.launch {
-//            val retrofit = provideRetrofit(provideOkHttpClient())
-//            val accountsApi: AccountsApi = retrofit.create(AccountsApi::class.java)
-//            println(accountsApi.getAccount().count())
-//        }
-
         setContent {
             MyFinanceTheme {
                 val navController = rememberNavController()
@@ -36,7 +30,7 @@ class MainActivity : ComponentActivity() {
                 ) { innerPadding ->
                     FinappNavHost(
                         navController,
-                        modifier = Modifier.Companion.padding(innerPadding)
+                        modifier = Modifier.padding(innerPadding)
                     )
                 }
             }
