--- app/src/main/java/com/example/myfinance/data/api/CategoriesApi.kt
@@ -1,13 +1,0 @@
-package com.example.myfinance.data.api
-
-import com.example.myfinance.data.dto.CategoryDto
-import retrofit2.http.GET
-import retrofit2.http.Path
-
-interface CategoriesApi {
-    @GET("categories")
-    suspend fun getCategories(): List<CategoryDto>
-
-    @GET("categories/type/{isIncome}")
-    suspend fun getCategoriesByType(@Path("isIncome") isIncome: Boolean): List<CategoryDto>
-}
