+++ app/src/main/java/com/example/myfinance/data/api/category/CategoryApi.kt
@@ -1,0 +1,14 @@
+package com.example.myfinance.data.api.category
+
+import com.example.myfinance.data.model.CategoryDto
+import retrofit2.Response
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface CategoryApi {
+    @GET("categories")
+    suspend fun getAllCategories(): Response<List<CategoryDto>>
+
+    @GET("categories/type/{isIncome}")
+    suspend fun getCategoriesByType(@Path("isIncome") isIncome: Boolean): Response<List<CategoryDto>>
+}
