--- app/src/main/java/com/example/myfinance/feature/domain/Account.kt
@@ -1,7 +1,0 @@
-package com.example.myfinance.feature.domain
-
-data class Account(
-    val id: Int,
-    val balance: Int,
-    val currency: String = "₽"
-)
