--- app/src/main/java/com/example/myfinance/feature/domain/Article.kt
@@ -1,7 +1,0 @@
-package com.example.myfinance.feature.domain
-
-data class Article(
-    val id: Int,
-    val name: String,
-    val emoji: String
-)
