--- app/src/main/java/com/example/myfinance/feature/domain/Income.kt
@@ -1,8 +1,0 @@
-package com.example.myfinance.feature.domain
-
-data class Income(
-    val id: Int,
-    val category: String,
-    val amount: Int,
-    val comment: String = ""
-)
