--- app/src/main/java/com/example/myfinance/feature/domain/Settings.kt
@@ -1,6 +1,0 @@
-package com.example.myfinance.feature.domain
-
-data class Settings(
-    val id: Int,
-    val name: String
-)
