+++ app/src/main/java/com/example/myfinance/feature/domain/model/Account.kt
@@ -1,0 +1,7 @@
+package com.example.myfinance.feature.domain.model
+
+data class Account(
+    val id: Int,
+    val balance: Int,
+    val currency: String = "₽"
+)
