+++ app/src/main/java/com/example/myfinance/feature/domain/model/Article.kt
@@ -1,0 +1,7 @@
+package com.example.myfinance.feature.domain.model
+
+data class Article(
+    val id: Int,
+    val name: String,
+    val emoji: String
+)
