+++ app/src/main/java/com/example/myfinance/feature/domain/model/Settings.kt
@@ -1,0 +1,6 @@
+package com.example.myfinance.feature.domain.model
+
+data class Settings(
+    val id: Int,
+    val name: String
+)
